Sun temp

+++ minishell_sw/env/env_utils2.c
@@ -6,7 +6,7 @@
 /*   By: sunwsong <sunwsong@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/30 13:43:15 by sunwsong          #+#    #+#             */
-/*   Updated: 2023/02/18 19:42:00 by sunwsong         ###   ########.fr       */
+/*   Updated: 2023/02/20 15:17:46 by sunwsong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -118,7 +118,7 @@ char	*find_env_val(char const *key)
 	return (NULL);
 }
 
-int	set_exitcode(int exit_code, int ret)
+long long	set_exitcode(int exit_code, long long ret)
 {
 	char	*code;
